+++ packages/cli/src/gdal/gdal.docker.ts
@@ -3,8 +3,8 @@ import * as os from 'os';
 import * as path from 'path';
 import { GdalCommand } from './gdal.command';
 
-const DOCKER_CONTAINER = Env.get(Env.Gdal.DockerContainer, 'osgeo/gdal');
-const DOCKER_CONTAINER_TAG = Env.get(Env.Gdal.DockerContainerTag, 'ubuntu-small-latest');
+const DOCKER_CONTAINER = Env.get(Env.Gdal.DockerContainer) ?? 'osgeo/gdal';
+const DOCKER_CONTAINER_TAG = Env.get(Env.Gdal.DockerContainerTag) ?? 'ubuntu-small-latest';
 
 export class GdalDocker extends GdalCommand {
     mounts: string[];
@@ -15,13 +15,10 @@ export class GdalDocker extends GdalCommand {
     }
 
     mount(filePath: string): void {
-        if (FileOperator.isS3(filePath)) {
-            return;
-        }
+        if (FileOperator.isS3(filePath)) return;
+
         const basePath = path.dirname(filePath);
-        if (this.mounts.includes(basePath)) {
-            return;
-        }
+        if (this.mounts.includes(basePath)) return;
         this.mounts.push(basePath);
     }