+++ packages/shared/src/proj/projection.ts
@@ -1,4 +1,5 @@
-import { BoundingBox, Epsg, EpsgCode, GeoJson } from '@basemaps/geo';
+import { BoundingBox, Bounds, Epsg, EpsgCode, GeoJson } from '@basemaps/geo';
+import { BBox, MultiPolygon, multiPolygonToWgs84 } from '@linzjs/geojson';
 import { Position } from 'geojson';
 import Proj from 'proj4';
 import { NamedBounds } from '../aws/tile.metadata.base';
@@ -14,7 +15,7 @@ const CodeMap = new Map<EpsgCode, Projection>();
 export class Projection {
     epsg: Epsg;
 
-    /** Transform coordinates to and from Wsg84 */
+    /** Transform coordinates to and from Wgs84 */
     private projection: proj4.Converter;
 
     /**
@@ -73,24 +74,38 @@ export class Projection {
     projectMultipolygon(multipoly: Position[][][], targetProjection: Projection): Position[][][] {
         if (targetProjection.epsg.code === this.epsg.code) return multipoly;
 
-        const { toWsg84 } = this;
-        const { fromWsg84 } = targetProjection;
+        const { toWgs84 } = this;
+        const { fromWgs84 } = targetProjection;
 
-        return multipoly.map((poly) => poly.map((ring) => ring.map((p) => fromWsg84(toWsg84(p)))));
+        return multipoly.map((poly) => poly.map((ring) => ring.map((p) => fromWgs84(toWgs84(p)))));
     }
 
     /**
      * Convert source `[x, y]` coordinates to `[lon, lat]`
      */
-    get toWsg84(): (coordinates: Position) => Position {
+    get toWgs84(): (coordinates: Position) => Position {
         return this.projection.forward;
     }
 
     /**
      * Convert `[lon, lat]` coordinates to source `[x, y]`
      */
-    get fromWsg84(): (coordinates: Position) => Position {
+    get fromWgs84(): (coordinates: Position) => Position {
         return this.projection.inverse;
+    }
+
+    /**
+     * Convert a source Bounds to GeoJSON WGS84 BoundingBox. In particular if the bounds crosses the
+     * anti-meridian then the east component will be less than the west component.
+
+     * @param source
+     * @returns [west, south, east, north]
+     */
+    boundsToWgs84BoundingBox(source: Bounds): BBox {
+        const sw = this.toWgs84([source.x, source.y]);
+        const ne = this.toWgs84([source.right, source.bottom]);
+
+        return [sw[0], sw[1], ne[0], ne[1]];
     }
 
     /**
@@ -98,30 +113,23 @@ export class Projection {
 
      * @param bounds in source epsg
      * @param properties any properties to include in the feature such as name
+
+     * @returns If `bounds` crosses the antimeridian then and east and west pair of non crossing
+     * polygons will be returned; otherwise a single Polygon will be returned.
      */
     boundsToGeoJsonFeature(bounds: BoundingBox, properties = {}): GeoJSON.Feature {
-        const { toWsg84 } = this;
-        const sw = toWsg84([bounds.x, bounds.y]);
-        const se = toWsg84([bounds.x + bounds.width, bounds.y]);
-        const nw = toWsg84([bounds.x, bounds.y + bounds.height]);
-        const ne = toWsg84([bounds.x + bounds.width, bounds.y + bounds.height]);
+        const sw = [bounds.x, bounds.y];
+        const se = [bounds.x + bounds.width, bounds.y];
+        const nw = [bounds.x, bounds.y + bounds.height];
+        const ne = [bounds.x + bounds.width, bounds.y + bounds.height];
 
-        if (sw[0] < se[0] && nw[0] < ne[0]) {
-            return GeoJson.toFeaturePolygon([[sw, nw, ne, se, sw]], properties);
+        const coords = multiPolygonToWgs84([[[sw, nw, ne, se, sw]]] as MultiPolygon, this.toWgs84);
+
+        if (coords.length == 1) {
+            return GeoJson.toFeaturePolygon(coords[0], properties);
         }
 
-        // crosses antimeridian so need to split polygon at antimeridian
-
-        // calculate where antimeridian is at north and south bounds
-        const northFraction = (180 - nw[0]) / (ne[0] + 360 - nw[0]);
-        const southFraction = (180 - sw[0]) / (se[0] + 360 - sw[0]);
-        const n180 = toWsg84([bounds.x + bounds.width * northFraction, bounds.y + bounds.height])[1];
-        const s180 = toWsg84([bounds.x + bounds.width * southFraction, bounds.y])[1];
-
-        return GeoJson.toFeatureMultiPolygon(
-            [[[sw, [180, s180], [180, n180], nw, sw]], [[se, ne, [-180, n180], [-180, s180], se]]],
-            properties,
-        );
+        return GeoJson.toFeatureMultiPolygon(coords, properties);
     }
 
     /** Convert a tile covering to a GeoJSON FeatureCollection */